Bound `since_id` from below with the current time
The old implementation used same `since_id`s until a new Tweet appears
in the list timeline, retrieving the same Tweet with the ID of that
`since_id`, regardless of how old it is.

The new implementation tries to skip retrieving such Tweets if they are
older than `twitter.delay` seconds.